frontend refactored, warnings added when registering an order and documentation added
