+++ static/css/graficos.css
@@ -1,0 +1,68 @@
+*, *::before, *::after { box-sizing: border-box; margin: 0; padding: 0; }
+:root {
+  --bg: #0f0e0b; --surface: #1a1916; --surface2: #222118;
+  --border: #2e2c25; --accent: #e8b84b; --accent2: #c9943a;
+  --text: #f0ece0; --muted: #7a776a; --danger: #c0392b;
+  --success: #4caf83; --radius: 10px;
+}
+html, body { height: 100%; }
+body { background: var(--bg); color: var(--text); font-family: 'Inter', sans-serif; height: 100vh; display: flex; flex-direction: column; overflow: hidden; }
+
+/* HEADER */
+header { flex-shrink: 0; padding: 14px 36px; border-bottom: 1px solid var(--border); display: flex; align-items: center; gap: 20px; background: var(--surface); }
+.logo-img { height: 52px; width: auto; }
+.logo-divider { width: 1px; height: 36px; background: var(--border); }
+.logo-text { display: flex; flex-direction: column; gap: 2px; }
+.logo-name { font-family: 'Syne', sans-serif; font-size: 17px; font-weight: 800; color: var(--accent); }
+.logo-sub { font-family: 'DM Mono', monospace; font-size: 10px; color: var(--muted); letter-spacing: 1.5px; text-transform: uppercase; }
+.page-title { font-family: 'Syne', sans-serif; font-size: 16px; font-weight: 800; color: var(--muted); }
+.btn-voltar { margin-left: auto; display: flex; align-items: center; gap: 7px; background: none; border: 1px solid var(--border); border-radius: 8px; color: var(--muted); font-family: 'Inter', sans-serif; font-size: 13px; padding: 7px 16px; cursor: pointer; text-decoration: none; transition: all 0.2s; }
+.btn-voltar:hover { border-color: var(--accent); color: var(--accent); }
+
+/* CONTENT */
+.content { flex: 1; overflow-y: auto; }
+.content::-webkit-scrollbar { width: 6px; }
+.content::-webkit-scrollbar-thumb { background: var(--border); border-radius: 3px; }
+.graficos-inner { display: grid; grid-template-columns: 1fr 1fr; min-height: 100%; }
+.graficos-inner .grafico-bloco:nth-child(n+3) { border-top: 1px solid var(--border); }
+.grafico-bloco { padding: 24px 32px 28px; background: var(--surface); }
+.grafico-bloco + .grafico-bloco { border-left: 1px solid var(--border); }
+.gb-titulo { font-family: 'Syne', sans-serif; font-size: 15px; font-weight: 800; margin-bottom: 14px; }
+
+/* FILTROS */
+.gb-filtros { display: flex; gap: 8px; align-items: flex-end; flex-wrap: wrap; margin-bottom: 16px; padding-bottom: 14px; border-bottom: 1px solid var(--border); }
+.gb-fi { display: flex; flex-direction: column; gap: 3px; position: relative; }
+.gb-fi-lbl { font-family: 'DM Mono', monospace; font-size: 9px; letter-spacing: 1.5px; text-transform: uppercase; color: var(--muted); }
+.gb-fi input[type="date"] { background: var(--surface2); border: 1px solid var(--border); border-radius: 7px; color: var(--text); font-family: 'Inter', sans-serif; font-size: 12px; padding: 5px 8px; outline: none; transition: border-color 0.2s; width: 136px; }
+.gb-fi input[type="date"]:focus { border-color: var(--accent); }
+.gb-fi input[type="date"]::-webkit-calendar-picker-indicator { filter: invert(0.6); cursor: pointer; }
+.gb-ms-trigger { display: flex; align-items: center; justify-content: space-between; gap: 4px; background: var(--surface2); border: 1px solid var(--border); border-radius: 7px; color: var(--muted); font-family: 'Inter', sans-serif; font-size: 12px; padding: 6px 8px 6px 10px; cursor: pointer; transition: border-color 0.2s; white-space: nowrap; text-align: left; }
+.gb-ms-trigger:hover { border-color: var(--accent); }
+.gb-ms-trigger.active { border-color: var(--accent); color: var(--accent); }
+.gb-ms-trigger.open { border-bottom-left-radius: 0; border-bottom-right-radius: 0; border-color: var(--accent); }
+.gb-ms-trigger.open .ms-arrow { transform: rotate(180deg); }
+.gb-w-produto { width: 140px; } .gb-w-tipo { width: 100px; } .gb-w-cliente { width: 140px; }
+.ms-arrow { font-size: 9px; color: var(--muted); flex-shrink: 0; transition: transform 0.15s; }
+.ms-trigger-text { overflow: hidden; text-overflow: ellipsis; flex: 1; }
+.gb-ms-panel { display: none; position: absolute; top: 100%; left: 0; min-width: 100%; background: var(--surface); border: 1px solid var(--accent); border-top: none; border-radius: 0 0 8px 8px; z-index: 500; box-shadow: 0 8px 24px rgba(0,0,0,0.55); flex-direction: column; max-height: 200px; }
+.gb-ms-panel.open { display: flex; }
+.ms-search-wrap { padding: 7px 8px 5px; border-bottom: 1px solid var(--border); flex-shrink: 0; }
+.ms-search { width: 100%; background: var(--surface2); border: 1px solid var(--border); border-radius: 5px; color: var(--text); font-size: 11px; font-family: 'Inter', sans-serif; padding: 4px 8px; outline: none; }
+.ms-search:focus { border-color: var(--accent); }
+.ms-list { overflow-y: auto; flex: 1; }
+.ms-list::-webkit-scrollbar { width: 4px; }
+.ms-list::-webkit-scrollbar-thumb { background: var(--border); border-radius: 2px; }
+.ms-opt { display: flex; align-items: center; gap: 8px; padding: 7px 10px; cursor: pointer; font-size: 12px; user-select: none; }
+.ms-opt:hover { background: rgba(232,184,75,0.08); }
+.ms-opt input[type="checkbox"] { accent-color: var(--accent); width: 13px; height: 13px; cursor: pointer; flex-shrink: 0; }
+.ms-opt span { overflow: hidden; text-overflow: ellipsis; white-space: nowrap; color: var(--text); }
+.ms-opt.hidden { display: none; }
+.ms-footer { padding: 5px 10px; border-top: 1px solid var(--border); display: flex; justify-content: space-between; align-items: center; flex-shrink: 0; }
+.ms-footer-btn { background: none; border: none; font-size: 11px; cursor: pointer; font-family: 'Inter', sans-serif; padding: 2px 4px; transition: color 0.15s; color: var(--muted); }
+.ms-footer-btn:hover { color: var(--accent); }
+.ms-footer-btn.danger:hover { color: var(--danger); }
+.gb-btn { background: none; border: 1px solid var(--border); border-radius: 7px; color: var(--muted); font-size: 11px; padding: 6px 10px; cursor: pointer; font-family: 'Inter', sans-serif; white-space: nowrap; transition: all 0.2s; height: 31px; }
+.gb-btn:hover { border-color: var(--accent); color: var(--accent); }
+.gb-btn.danger:hover { border-color: var(--danger); color: var(--danger); }
+.gb-canvas-wrap { position: relative; height: 210px; }
+.gb-empty { text-align: center; padding: 40px 0; color: var(--muted); font-size: 13px; display: none; }
